passport-photo-ai/deployment_backup_251226_114802/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
rembg==2.0.50
pillow-heif==0.13.0
boto3==1.34.0
botocore==1.34.0
python-dotenv==1.0.0
onnxruntime==1.16.3
mediapipe==0.10.9
hypothesis==6.92.1
# Additional dependencies for enhanced processing
scikit-image==0.21.0
scipy==1.11.4
# For better AWS compatibility
setuptools==68.2.2
wheel==0.41.2